+++ Expense-tracking-System-backend/Expense-tracking-backend-main/user-service/src/main/java/com/jaya/task/user/service/service/UserService.java
@@ -18,7 +18,7 @@ public interface UserService {
 
     public User updateUserProfile(String jwt, UserUpdateRequest updateRequest);
 
-    void  deleteUser(Long id) throws AccessDeniedException;
+    void  deleteUser(Integer id) throws AccessDeniedException;
 
     public boolean checkEmailAvailability(String email);
